% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreForecast.R
\name{mse}
\alias{mse}
\title{Mean squared error}
\usage{
mse(observed, predicted)
}
\arguments{
\item{observed}{numeric vector of realized values}

\item{predicted}{numeric vector of point predictions (same length/order)}
}
\value{
a single numeric value
}
\description{
Mean squared error
}
\examples{
mse(c(10, 20, 30), c(12, 18, 33))
}
